adds an unfinished unit test for registering a tuya remote
Adds the start of a unit test. Needs the TuyaOpenAPI.connect method mocked. Attempted to copy the mock setup from test_config_flow and messed something up causing the unit tests not to load in the vscode explorer. Have commented the mocking code out for now to leave things in a "working" state for this commit